+++ Controller/GooglePay/Payment.php
@@ -1,0 +1,26 @@
+<?php
+namespace Monri\Payments\Controller\GooglePay;
+
+use Magento\Framework\App\Action\HttpGetActionInterface;
+use Magento\Framework\View\Result\PageFactory;
+
+class Payment implements HttpGetActionInterface
+{
+    /**
+     * Payment constructor.
+     *
+     * @param PageFactory $resultPageFactory
+     */
+    public function __construct(
+        private PageFactory $resultPageFactory
+    ) {
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function execute()
+    {
+        return $this->resultPageFactory->create();
+    }
+}