+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.7.0
 - Improved Monri WebPay tokenization by using API for transaction creation.
 - Added "Instant Purchase" on the product page for customers with saved Monri WebPay tokens.
+- Added option to disable installments for Monri WebPay.
 
 ### 1.6.0
 - WebPay tokenization (Beta)